[refactor] KafkaTopicService로 로직분리
